scripts/requirements.txt
# Requirements for Property Scraper and Data Import
requests>=2.31.0
supabase>=2.3.0
python-dotenv>=1.0.0
